+++ scripts/api-compare/extract-ts-api.ts
@@ -610,6 +610,17 @@ export function extractFromProgram(
         const name = node.name.text;
         const modKey = `${relPath}:${name}`;
         const extracted = extractNamespace(node, checker, relPath);
+        // A namespace may hold a nested `export class` (Ruby's `class` inside a
+        // `module`); register it as a class of its own so its members are seen.
+        if (node.body && ts.isModuleBlock(node.body)) {
+          for (const stmt of node.body.statements) {
+            if (!ts.isClassDeclaration(stmt) || !stmt.name || !isExported(stmt)) continue;
+            const nested = extractClass(stmt, checker, relPath, srcDir);
+            if (nested) {
+              info.classes[`${relPath}:${nested.name}`] = nested;
+            }
+          }
+        }
         const existing = info.modules[modKey];
         if (existing) {
           const existingNames = new Set(existing.instanceMethods.map((m) => m.name));
